refactor: update credit batches fetching
