+++ web-marketplace/src/components/organisms/CreditTotals/CreditTotals.types.ts
@@ -1,0 +1,5 @@
+export interface CreditTotalData {
+  tradeable?: string;
+  retired?: string;
+  created?: string;
+}
